LayoutTests/editing/selection/select-out-of-floated-non-editable-11.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div>block-level text before floats</div><div id="target" style="float: left">first float*</div><div style="float: left">second float</div></div>
<p style="padding-top: 1em;">
To test manually, drag from the middle of the 'first-float' div to the
right, into the the 'second float' element. The selection should go into
the 'second float' element, including text of both floats, and not jump
to the beginning.
</p>
<pre id="console"></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-out-of-floated-non-editable.js"></script>
<script>
runTest("float*s", 5);
</script>
</body>
</html>
