LayoutTests/editing/selection/select-out-of-floated-non-editable-11-expected.txt

block-level text before floats
first float*second float
To test manually, drag from the middle of the 'first-float' div to the right, into the the 'second float' element. The selection should go into the 'second float' element, including text of both floats, and not jump to the beginning.

SUCCESS: Selected test is 'float*s' as expected
| <div>
|   "block-level text before floats"
| <div>
|   id="target"
|   style="float: left"
|   "first <#selection-anchor>float*"
| <div>
|   style="float: left"
|   "s<#selection-focus>econd float"
